Add update-gh-pages.sh
This script iterates over all git clones in the current dir;
it then tests each for an origin/gh-pages branch, and a local
branch gh-pages tracking that remote branch.
It also adds a worktree for that tree, if missing.

Finally, it checks if that worktree resp. the gh-pages
branch is based on GitHubPagesForGAP, and in that case, attempts
to upgrade it to use the latest version of GitHubPagesForGAP